Move proto conversion tests to its own file
